pyproject.toml
[project]
name = "canteenkart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "autopep8>=2.3.2",
    "firebase-admin>=7.1.0",
    "flask>=3.1.2",
    "flask-login>=0.6.3",
    "flask-socketio>=5.5.1",
    "flask-sqlalchemy>=3.1.1",
    "pillow>=11.3.0",
    "pip>=25.2",
    "pytest>=8.4.2",
    "qrcode[pil]>=8.2",
]
